fixed page tables, switched bootlaoder to g++

+++ bootloader/include/drive.hpp
@@ -4,6 +4,6 @@
  *                          the BIOS hard drive driver
  */
 namespace drive {
-    extern "C" unsigned short read(void *dest, unsigned int segment, unsigned int sector, unsigned int amt);
-    extern "C" unsigned short write(void *src, unsigned int sector, unsigned char amount);
+    extern "C" unsigned short read(void *dest, unsigned short segment, unsigned short sector, unsigned short amt);
+    extern "C" unsigned short write(void *src, unsigned short sector, unsigned char amount);
 }  // namespace drive
